+++ authority-portal-frontend/src/app/shared/common/logout-button/logout-button.component.ts
@@ -18,5 +18,5 @@ import {APP_CONFIG, AppConfig} from 'src/app/core/services/config/app-config';
   templateUrl: './logout-button.component.html',
 })
 export class LogoutButtonComponent {
-  constructor(@Inject(APP_CONFIG) public config: AppConfig) {}
+  constructor(@Inject(APP_CONFIG) public appConfig: AppConfig) {}
 }